Permitir teléfonos largos en ciudadanos

+++ ciudadanos/models.py
@@ -694,8 +694,8 @@ class Ciudadano(models.Model):
     codigo_postal = models.IntegerField(
         null=True, blank=True, verbose_name="Código Postal"
     )
-    telefono = models.IntegerField(null=True, blank=True)
-    telefono_alternativo = models.IntegerField(
+    telefono = models.BigIntegerField(null=True, blank=True)
+    telefono_alternativo = models.BigIntegerField(
         null=True, blank=True, verbose_name="Telefono Alternativo"
     )
     email = models.EmailField(null=True, blank=True)
